river.py: add --initial flag
When ran with this flag, only entries published within the past 24
hours are included in the river.

This gives more feeds a chance to get onto the river and prevents
feeds with a lot of entries from drowing all others out.

Eventually I'd like this 'only show recent' behavior to kick in
whenever a new feed URL is discovered and to have the number of hours
configurable.