download.py: clean entry and title text before comparing
Turns out not all Github commits are guilty of duplicating title and
description. Only when a commit has no "body" and is all "title" are
the two fields the same.

But Github does add some HTML formatting to the "description" element,
so by cleaning it beforehand we get a better idea if they're "the same."