custom.js: clear expired setInterval calls
What was happening is I would change the river, but a previously
executed setInterval was still alive. Then when it fired, it would get
all confused because the river had changed.

Now, store the current interval in a global variable -- yeah, yeah, I
know -- and clear it when changing feeds.